test-utils: Add a PostgreSQL test listener
Add a `TestListener` that manages an instance of `EmbeddedPostgres` to
reduce code duplication.
