+++ scanner/build.gradle.kts
@@ -24,7 +24,6 @@ val jacksonVersion: String by project
 val kotlinxCoroutinesVersion: String by project
 val mockkVersion: String by project
 val postgresVersion: String by project
-val postgresEmbeddedVersion: String by project
 val retrofitVersion: String by project
 val scancodeVersion: String by project
 val sw360ClientVersion: String by project
@@ -68,8 +67,6 @@ dependencies {
 
     testImplementation("com.github.tomakehurst:wiremock:$wiremockVersion")
     testImplementation("io.mockk:mockk:$mockkVersion")
-
-    funTestImplementation("com.opentable.components:otj-pg-embedded:$postgresEmbeddedVersion")
 }
 
 buildConfig {
